+++ src/modules/files-storage/domain/factory/file-dto.factory.ts
@@ -1,6 +1,6 @@
 import { Readable } from 'node:stream';
 import { FileDto } from '../dto';
-import { StorageType } from '../storage-paths';
+import { StorageType } from '../storage-paths.const';
 
 export class FileDtoFactory {
 	public static create(
