+++ src/modules/files-storage/domain/factory/file-record.factory.ts
@@ -2,7 +2,7 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { EntityId } from '@shared/domain/types';
 import { FileRecord, FileRecordProps } from '../file-record.do';
 import { ParentInfo } from '../interface';
-import { StorageType } from '../storage-paths';
+import { StorageType } from '../storage-paths.const';
 import { FileRecordSecurityCheck } from '../vo';
 
 export class FileRecordFactory {
