+++ src/modules/files-storage/domain/index.ts
@@ -12,7 +12,7 @@ export * from './interface';
 export { FileStorageActionsLoggable } from './loggable';
 export { FilesStorageMapper } from './mapper';
 export { FilesStorageService, PreviewService } from './service';
-export * from './storage-paths';
+export * from './storage-paths.const';
 export {
 	FileRecordSecurityCheck,
 	FileRecordSecurityCheckProps,
