+++ src/modules/files-storage/domain/storage-paths.const.ts
